% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePercentDate_Beta_2_1.R
\name{f.ConvertPercentageDate}
\alias{f.ConvertPercentageDate}
\title{Changes the DateColumn from WUFI+ Output Files to a usable Format IF it is provided in percentages (Dont ask how this happened)}
\usage{
f.ConvertPercentageDate(DateColumn)
}
\arguments{
\item{DateColumn}{DateColumn of WUFI+ Output files.}
}
\value{
Date. DateCol of WUFIPlus in correct format
}
\description{
Changes the DateColumn from WUFI+ Output Files to a usable Format IF it is provided in percentages (Dont ask how this happened)
}
\examples{
x <- 1
}
