% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWUFIplusData_readr_sepanddate_Beta_2_2.r
\name{f.GetWufiPlusData_readr_b}
\alias{f.GetWufiPlusData_readr_b}
\title{Read WUFIPlus Data into R.}
\usage{
f.GetWufiPlusData_readr_b(v.Path, select.columns = c())
}
\arguments{
\item{v.Path}{Character. Path to file.}

\item{select.columns}{Vector. Which columns should be chosen? Used for timesaving, but readr eliminated that problem.}
}
\value{
Data. Tibble of WUFIPlus Data.
}
\description{
Read WUFIPlus Data into R.
}
