% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessDateforGetWUFIPlus_Beta_2_0.R
\name{f.GuessDate_for_GetWUFIplus}
\alias{f.GuessDate_for_GetWUFIplus}
\title{Guesses the Date format for the WUFIPlus DateColumn. Its pretty good at it.}
\usage{
f.GuessDate_for_GetWUFIplus(DateColumn)
}
\arguments{
\item{DateColumn}{WUFIPlus Datecolumn}
}
\value{
Character. v.Dateoformat is the string section for correct formatting of dates.
}
\description{
Guesses the Date format for the WUFIPlus DateColumn. Its pretty good at it.
}
\examples{
x <- 1
}
