% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDateColOnly_1_0.R
\name{ReadDateCol.Only}
\alias{ReadDateCol.Only}
\title{Old Testing function to only import the DateCol of a WUFIPlus file.}
\usage{
ReadDateCol.Only(v.Path)
}
\arguments{
\item{v.Path}{Character. Path to WUFIPlus file.}
}
\value{
Character. DateCol of WUFIPlus
}
\description{
Old Testing function to only import the DateCol of a WUFIPlus file.
}
\examples{
x <- 1
}
