% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrLft.R
\name{substrLeft}
\alias{substrLeft}
\title{Subtract n elements from a string x from the left side. Save the subtracted elements.}
\usage{
substrLeft(x, n)
}
\arguments{
\item{x}{Chracter. A string of characters}

\item{n}{Numeric. Amount of elements to be subtracted.}
}
\value{
Character. The n first characters of the string x.
}
\description{
Subtract n elements from a string x from the left side. Save the subtracted elements.
}
\examples{
substrLeft("WUFI + R = eternal love", 6)
}
